Update to 64bit game version. 32bit support dropped

+++ console.h
@@ -8,7 +8,7 @@ void console_log(const char* text);
 void console_confirm(const char* text);
 void console_log_1num(const char* text, int a);
 void console_log_1str(const char* text, const char* a);
-void console_log_1num_hex(const char* text, int a);
+void console_log_1num_hex(const char* text, LONGLONG a);
 void console_log_2num(const char* text, int a, int b);
 void console_print_char_stats();
 int console_get_input();
